+++ test-app/mocks/mockClients/dashboards/definitions/age-breakdown/vis-definitions.js
@@ -514,7 +514,6 @@ const mockBarChartReligion = {
         axis: 'y',
       },
     ],
-    ignoreRemainingColumns: true,
   },
 }
